fix: more re-organization and fixes after rebase
