+++ api/src/main/java/io/javaoperatorsdk/operator/api/Context.java
@@ -1,8 +1,6 @@
 package io.javaoperatorsdk.operator.api;
 
 import io.fabric8.kubernetes.client.CustomResource;
-import io.javaoperatorsdk.operator.processing.event.EventList;
-import io.javaoperatorsdk.operator.processing.event.EventSourceManager;
 import java.util.Optional;
 
 public interface Context<T extends CustomResource> {
