+++ api/src/main/java/io/javaoperatorsdk/operator/api/DeleteControl.java
@@ -1,6 +1,6 @@
 package io.javaoperatorsdk.operator.api;
 
 public enum DeleteControl {
-  DEFAULT_DELETE,
-  NO_FINALIZER_REMOVAL
+    DEFAULT_DELETE,
+    NO_FINALIZER_REMOVAL
 }
