+++ api/src/main/java/io/javaoperatorsdk/operator/api/Event.java
@@ -1,4 +1,4 @@
-package io.javaoperatorsdk.operator.processing.event;
+package io.javaoperatorsdk.operator.api;
 
 public interface Event {
