+++ api/src/main/java/io/javaoperatorsdk/operator/api/EventHandler.java
@@ -1,0 +1,7 @@
+package io.javaoperatorsdk.operator.api;
+
+public interface EventHandler {
+    
+    void handleEvent(Event event);
+    
+}
