+++ api/src/main/java/io/javaoperatorsdk/operator/api/EventList.java
@@ -1,20 +1,20 @@
-package io.javaoperatorsdk.operator.processing.event;
+package io.javaoperatorsdk.operator.api;
 
 import java.util.List;
 import java.util.Optional;
 
 public class EventList {
-
+  
   private final List<Event> eventList;
-
+  
   public EventList(List<Event> eventList) {
     this.eventList = eventList;
   }
-
+  
   public List<Event> getList() {
     return eventList;
   }
-
+  
   public <T extends Event> Optional<T> getLatestOfType(Class<T> eventType) {
     for (int i = eventList.size() - 1; i >= 0; i--) {
       Event event = eventList.get(i);
