+++ api/src/main/java/io/javaoperatorsdk/operator/api/EventSource.java
@@ -1,0 +1,9 @@
+package io.javaoperatorsdk.operator.api;
+
+public interface EventSource {
+    
+    void setEventHandler(EventHandler eventHandler);
+    
+    void eventSourceDeRegisteredForResource(String customResourceUid);
+    
+}
