+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/EventListUtils.java
@@ -1,10 +1,10 @@
 package io.javaoperatorsdk.operator;
 
-import io.fabric8.kubernetes.client.Watcher;
-import io.javaoperatorsdk.operator.processing.event.Event;
-import io.javaoperatorsdk.operator.processing.event.internal.CustomResourceEvent;
 import java.util.List;
 
+import io.fabric8.kubernetes.client.Watcher;
+import io.javaoperatorsdk.operator.api.Event;
+import io.javaoperatorsdk.operator.processing.event.internal.CustomResourceEvent;
 public class EventListUtils {
 
   public static boolean containsCustomResourceDeletedEvent(List<Event> events) {
