+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/DefaultEventHandler.java
@@ -4,22 +4,23 @@ import static io.javaoperatorsdk.operator.EventListUtils.containsCustomResourceD
 import static io.javaoperatorsdk.operator.processing.KubernetesResourceUtils.getUID;
 import static io.javaoperatorsdk.operator.processing.KubernetesResourceUtils.getVersion;
 
-import io.fabric8.kubernetes.client.CustomResource;
-import io.javaoperatorsdk.operator.api.RetryInfo;
-import io.javaoperatorsdk.operator.processing.event.DefaultEventSourceManager;
-import io.javaoperatorsdk.operator.processing.event.Event;
-import io.javaoperatorsdk.operator.processing.event.EventHandler;
-import io.javaoperatorsdk.operator.processing.retry.Retry;
-import io.javaoperatorsdk.operator.processing.retry.RetryExecution;
-import java.util.*;
 import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.locks.ReentrantLock;
+
+import io.fabric8.kubernetes.client.CustomResource;
+import io.javaoperatorsdk.operator.api.Event;
+import io.javaoperatorsdk.operator.api.EventHandler;
+import io.javaoperatorsdk.operator.processing.event.DefaultEventSourceManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import static io.javaoperatorsdk.operator.EventListUtils.containsCustomResourceDeletedEvent;
+import static io.javaoperatorsdk.operator.processing.KubernetesResourceUtils.getUID;
+import static io.javaoperatorsdk.operator.processing.KubernetesResourceUtils.getVersion;
 
 /**
  * Event handler that makes sure that events are processed in a "single threaded" way per resource
