+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/EventBuffer.java
@@ -1,6 +1,6 @@
 package io.javaoperatorsdk.operator.processing;
 
-import io.javaoperatorsdk.operator.processing.event.Event;
+import io.javaoperatorsdk.operator.api.Event;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
