+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/ExecutionScope.java
@@ -1,16 +1,15 @@
 package io.javaoperatorsdk.operator.processing;
 
-import io.fabric8.kubernetes.client.CustomResource;
-import io.javaoperatorsdk.operator.api.RetryInfo;
-import io.javaoperatorsdk.operator.processing.event.Event;
 import java.util.List;
+
+import io.fabric8.kubernetes.client.CustomResource;
+import io.javaoperatorsdk.operator.api.Event;
 
 public class ExecutionScope {
 
   private List<Event> events;
   // the latest custom resource from cache
   private CustomResource customResource;
-
   private RetryInfo retryInfo;
 
   public ExecutionScope(List<Event> list, CustomResource customResource, RetryInfo retryInfo) {
