+++ operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/AbstractEvent.java
@@ -1,11 +1,14 @@
 package io.javaoperatorsdk.operator.processing.event;
 
+import io.javaoperatorsdk.operator.api.Event;
+import io.javaoperatorsdk.operator.api.EventSource;
+
 public abstract class AbstractEvent implements Event {
-
+  
   private final String relatedCustomResourceUid;
-
+  
   private final EventSource eventSource;
-
+  
   public AbstractEvent(String relatedCustomResourceUid, EventSource eventSource) {
     this.relatedCustomResourceUid = relatedCustomResourceUid;
     this.eventSource = eventSource;
