--- operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/EventHandler.java
@@ -1,6 +1,0 @@
-package io.javaoperatorsdk.operator.processing.event;
-
-public interface EventHandler {
-
-  void handleEvent(Event event);
-}
