--- operator-framework/src/main/java/io/javaoperatorsdk/operator/processing/event/EventSource.java
@@ -1,8 +1,0 @@
-package io.javaoperatorsdk.operator.processing.event;
-
-public interface EventSource {
-
-  void setEventHandler(EventHandler eventHandler);
-
-  void eventSourceDeRegisteredForResource(String customResourceUid);
-}
