+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/processing/event/EventListTest.java
@@ -1,5 +1,12 @@
 package io.javaoperatorsdk.operator.processing.event;
 
+import java.util.Arrays;
+
+import io.javaoperatorsdk.operator.api.Event;
+import io.javaoperatorsdk.operator.api.EventList;
+import io.javaoperatorsdk.operator.processing.event.internal.TimerEvent;
+import org.junit.jupiter.api.Test;
+
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.Mockito.mock;
