+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/processing/event/internal/TimerEventSourceTest.java
@@ -12,7 +12,7 @@ import static org.mockito.Mockito.verify;
 import io.fabric8.kubernetes.client.CustomResource;
 import io.javaoperatorsdk.operator.TestUtils;
 import io.javaoperatorsdk.operator.processing.KubernetesResourceUtils;
-import io.javaoperatorsdk.operator.processing.event.EventHandler;
+import io.javaoperatorsdk.operator.api.EventHandler;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
